backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
pyjwt==2.6.0
moralis==0.1.27
